# BackendSpring/src/main/java/heg/backendspring/api/StudentApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package heg.backendspring.api;

import heg.backendspring.models.StudentDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-23T19:52:08.287321300+01:00[Europe/Zurich]", comments = "Generator version: 7.9.0")
@Validated
@Tag(name = "student", description = "Gestion des étudiants liés aux cours")
public interface StudentApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /course/{courseId}/addStudent : Ajout d’un étudiant (route spécifique).
     * Ajoute un étudiant à un cours via une route spécifique existante.
     *
     * @param courseId Identifiant du cours. (required)
     * @param studentDto Données de l’étudiant à ajouter. (required)
     * @return Étudiant ajouté. (status code 201)
     *         or Cours introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "addStudentLegacy",
        summary = "Ajout d’un étudiant (route spécifique).",
        description = "Ajoute un étudiant à un cours via une route spécifique existante.",
        tags = { "student" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Étudiant ajouté.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = heg.backendspring.models.StudentDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Cours introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/course/{courseId}/addStudent",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<heg.backendspring.models.StudentDto> addStudentLegacy(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId,
        @Parameter(name = "StudentDto", description = "Données de l’étudiant à ajouter.", required = true) @Valid @RequestBody StudentDto studentDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 300, \"name\" : \"George Dylan\", \"email\" : \"george.dylan@etu.hesge.ch\", \"studyType\" : \"FULL_TIME\", \"studentCourses\" : [ { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /course/{courseId}/students : Ajout d’un étudiant à un cours.
     * Ajoute un étudiant à un cours via son DTO.
     *
     * @param courseId Identifiant du cours. (required)
     * @param studentDto Données de l’étudiant à ajouter. (required)
     * @return Étudiant ajouté au cours. (status code 201)
     *         or Cours introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "addStudentToCourse",
        summary = "Ajout d’un étudiant à un cours.",
        description = "Ajoute un étudiant à un cours via son DTO.",
        tags = { "student" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Étudiant ajouté au cours.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = heg.backendspring.models.StudentDto.class))
            }),
            @ApiResponse(responseCode = "404", description = "Cours introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/course/{courseId}/students",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<heg.backendspring.models.StudentDto> addStudentToCourse(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId,
        @Parameter(name = "StudentDto", description = "Données de l’étudiant à ajouter.", required = true) @Valid @RequestBody StudentDto studentDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 300, \"name\" : \"George Dylan\", \"email\" : \"george.dylan@etu.hesge.ch\", \"studyType\" : \"FULL_TIME\", \"studentCourses\" : [ { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /course/{courseId}/addStudentsFromFile : Ajout d’étudiants depuis un fichier.
     * Ajoute plusieurs étudiants à partir d’un fichier texte.
     *
     * @param courseId Identifiant du cours. (required)
     * @param file  (optional)
     * @return Étudiants ajoutés. (status code 200)
     *         or Fichier invalide. (status code 400)
     *         or Cours introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "addStudentsFromFile",
        summary = "Ajout d’étudiants depuis un fichier.",
        description = "Ajoute plusieurs étudiants à partir d’un fichier texte.",
        tags = { "student" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Étudiants ajoutés.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = heg.backendspring.models.StudentDto.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Fichier invalide."),
            @ApiResponse(responseCode = "404", description = "Cours introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/course/{courseId}/addStudentsFromFile",
        produces = { "application/json" },
        consumes = { "multipart/form-data" }
    )
    
    default ResponseEntity<List<heg.backendspring.models.StudentDto>> addStudentsFromFile(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId,
        @Parameter(name = "file", description = "") @RequestPart(value = "file", required = false) MultipartFile file
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : 300, \"name\" : \"George Dylan\", \"email\" : \"george.dylan@etu.hesge.ch\", \"studyType\" : \"FULL_TIME\", \"studentCourses\" : [ { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\" } ] }, { \"id\" : 300, \"name\" : \"George Dylan\", \"email\" : \"george.dylan@etu.hesge.ch\", \"studyType\" : \"FULL_TIME\", \"studentCourses\" : [ { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\" } ] } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /course/{courseId}/students : Liste des étudiants d’un cours.
     * Retourne les étudiants inscrits à un cours.
     *
     * @param courseId Identifiant du cours. (required)
     * @return Liste d’étudiants retournée. (status code 200)
     *         or Cours introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "getStudentsByCourse",
        summary = "Liste des étudiants d’un cours.",
        description = "Retourne les étudiants inscrits à un cours.",
        tags = { "student" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste d’étudiants retournée.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = heg.backendspring.models.StudentDto.class)))
            }),
            @ApiResponse(responseCode = "404", description = "Cours introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/course/{courseId}/students",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<heg.backendspring.models.StudentDto>> getStudentsByCourse(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : 300, \"name\" : \"George Dylan\", \"email\" : \"george.dylan@etu.hesge.ch\", \"studyType\" : \"FULL_TIME\", \"studentCourses\" : [ { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\" } ] }, { \"id\" : 300, \"name\" : \"George Dylan\", \"email\" : \"george.dylan@etu.hesge.ch\", \"studyType\" : \"FULL_TIME\", \"studentCourses\" : [ { \"id\" : 1, \"name\" : \"Python Introduction\", \"code\" : \"61-13\", \"semester\" : \"AUTOMNE\", \"courseType\" : \"PYTHON\", \"year_course\" : 2025, \"teacher\" : \"Stettler\" } ] } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /course/{courseId}/students/{id} : Suppression d’un étudiant d’un cours.
     * Retire un étudiant d’un cours à partir de son identifiant.
     *
     * @param courseId Identifiant du cours. (required)
     * @param id Identifiant de l’étudiant. (required)
     * @return Étudiant retiré du cours. (status code 204)
     *         or Cours ou étudiant introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "removeStudentFromCourse",
        summary = "Suppression d’un étudiant d’un cours.",
        description = "Retire un étudiant d’un cours à partir de son identifiant.",
        tags = { "student" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Étudiant retiré du cours."),
            @ApiResponse(responseCode = "404", description = "Cours ou étudiant introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/course/{courseId}/students/{id}"
    )
    
    default ResponseEntity<Void> removeStudentFromCourse(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId,
        @Parameter(name = "id", description = "Identifiant de l’étudiant.", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /course/{courseId}/removeStudent/{id} : Suppression d’un étudiant (route spécifique).
     * Supprime un étudiant d’un cours via une route spécifique existante.
     *
     * @param courseId Identifiant du cours. (required)
     * @param id Identifiant de l’étudiant. (required)
     * @return Étudiant supprimé. (status code 204)
     *         or Cours ou étudiant introuvable. (status code 404)
     *         or Erreur interne. (status code 500)
     */
    @Operation(
        operationId = "removeStudentLegacy",
        summary = "Suppression d’un étudiant (route spécifique).",
        description = "Supprime un étudiant d’un cours via une route spécifique existante.",
        tags = { "student" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Étudiant supprimé."),
            @ApiResponse(responseCode = "404", description = "Cours ou étudiant introuvable."),
            @ApiResponse(responseCode = "500", description = "Erreur interne.")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/course/{courseId}/removeStudent/{id}"
    )
    
    default ResponseEntity<Void> removeStudentLegacy(
        @Parameter(name = "courseId", description = "Identifiant du cours.", required = true, in = ParameterIn.PATH) @PathVariable("courseId") Long courseId,
        @Parameter(name = "id", description = "Identifiant de l’étudiant.", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}