# BackendSpring/src/main/resources/openapi/openapi.yaml
...
openapi: 3.0.4
info:
  title: AssistantIO API
  description: API de gestion des cours, TPs, étudiants et rendus.
  version: 1.0.0

# ==============================
#   SERVERS
# ==============================
servers:
  - url: http://localhost:8080
    description: Environnement local

# ==============================
#   TAGS
# ==============================
...
paths:
...
  /course/{courseId}/students:
    get:
      tags: [student]
      summary: Liste des étudiants d’un cours.
      description: Retourne les étudiants inscrits à un cours.
      operationId: getStudentsByCourse
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Liste d’étudiants retournée.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentDto'
        '404':
          description: Cours introuvable.
        '500':
          description: Erreur interne.
    post:
      tags: [student]
      summary: Ajout d’un étudiant à un cours.
      description: Ajoute un étudiant à un cours via son DTO.
      operationId: addStudentToCourse
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
      requestBody:
        description: Données de l’étudiant à ajouter.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
      responses:
        '201':
          description: Étudiant ajouté au cours.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDto'
        '404':
          description: Cours introuvable.
        '500':
          description: Erreur interne.
  /course/{courseId}/students/{id}:
    delete:
      tags: [student]
      summary: Suppression d’un étudiant d’un cours.
      description: Retire un étudiant d’un cours à partir de son identifiant.
      operationId: removeStudentFromCourse
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          description: Identifiant de l’étudiant.
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Étudiant retiré du cours.
        '404':
          description: Cours ou étudiant introuvable.
        '500':
          description: Erreur interne.

  # ========================================
  #   STUDENTS D’UN COURS (ROUTES SPÉCIFIQUES EXISTANTES)
  # ========================================
  /course/{courseId}/addStudent:
    post:
      tags: [student]
      summary: Ajout d’un étudiant (route spécifique).
      description: Ajoute un étudiant à un cours via une route spécifique existante.
      operationId: addStudentLegacy
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
      requestBody:
        description: Données de l’étudiant à ajouter.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentDto'
      responses:
        '201':
          description: Étudiant ajouté.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentDto'
        '404':
          description: Cours introuvable.
        '500':
          description: Erreur interne.
  /course/{courseId}/removeStudent/{id}:
    delete:
      tags: [student]
      summary: Suppression d’un étudiant (route spécifique).
      description: Supprime un étudiant d’un cours via une route spécifique existante.
      operationId: removeStudentLegacy
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
        - name: id
          in: path
          required: true
          description: Identifiant de l’étudiant.
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Étudiant supprimé.
        '404':
          description: Cours ou étudiant introuvable.
        '500':
          description: Erreur interne.
  /course/{courseId}/addStudentsFromFile:
    post:
      tags: [student]
      summary: Ajout d’étudiants depuis un fichier.
      description: Ajoute plusieurs étudiants à partir d’un fichier texte.
      operationId: addStudentsFromFile
      parameters:
        - name: courseId
          in: path
          required: true
          description: Identifiant du cours.
          schema:
            type: integer
            format: int64
      requestBody:
        description: Fichier texte contenant la liste des étudiants.
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Étudiants ajoutés.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentDto'
        '400':
          description: Fichier invalide.
        '404':
          description: Cours introuvable.
        '500':
          description: Erreur interne.

  # ========================================
  #   TPS D’UN COURS (REST PRINCIPAL)
  # ========================================
...
components:
  schemas:
...
    SemesterType:
      type: string
      description: Semestre académique.
      example: "AUTOMNE"
    CourseType:
      type: string
      description: Type de cours.
      example: "PYTHON"
    StudyType:
      type: string
      description: Type d’étude.
      example: "FULL_TIME"

    # --- DTO principaux exposés ---
    CourseDto:
      type: object
      description: Représentation d’un cours.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Python Introduction"
        code:
          type: string
          example: "61-13"
        semester:
          $ref: '#/components/schemas/SemesterType'
        courseType:
          $ref: '#/components/schemas/CourseType'
        year_course:
          type: integer
          example: 2025
        teacher:
          type: string
          example: "Stettler"
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentDto'
        tps:
          type: array
          items:
            $ref: '#/components/schemas/TPDto'
        evaluations:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationDto'
      required: ["id", "name", "code"]
      example:
        id: 1
        name: "Python Introduction"
        code: "61-13"
        semester: "AUTOMNE"
        courseType: "PYTHON"
        year_course: 2025
        teacher: "Stettler"
        students: []
        tps:
          - id: 10
            no: 1
            courseId: 1
            submission:
              id: 100
              fileName: "TP1_Rendus.zip"
              pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
              pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
            statusStudents: []
        evaluations:
          - id: 200
            name: "Examen final"
            date: "2025-01-15"
            coefficient: 2
            courseId: 1
    EvaluationDto:
      type: object
      description: Représentation d’une évaluation.
      properties:
        id:
          type: integer
          format: int64
          example: 200
        name:
          type: string
          example: "Examen final"
        date:
          type: string
          format: date
          example: "2025-01-15"
        coefficient:
          type: integer
          example: 2
        courseId:
          type: integer
          format: int64
          example: 1
      required: ["name", "date", "coefficient", "courseId"]
      example:
        id: 200
        name: "Examen final"
        date: "2025-01-15"
        coefficient: 2
        courseId: 1
    StudentDto:
      type: object
      description: Représentation d’un étudiant.
      properties:
        id:
          type: integer
          format: int64
          example: 300
        name:
          type: string
          example: "George Dylan"
        email:
          type: string
          format: email
          example: "george.dylan@etu.hesge.ch"
        studyType:
          $ref: '#/components/schemas/StudyType'
        studentCourses:
          type: array
          items:
            $ref: '#/components/schemas/CourseDto'
      required: ["name", "email", "studyType"]
      example:
        id: 300
        name: "George Dylan"
        email: "george.dylan@etu.hesge.ch"
        studyType: "FULL_TIME"
        studentCourses:
          - id: 1
            name: "Python Introduction"
            code: "61-13"
            semester: "AUTOMNE"
            courseType: "PYTHON"
            year_course: 2025
            teacher: "Stettler"
    SubmissionDto:
      type: object
      description: Représentation d’un rendu (fichier ZIP) pour un TP.
      properties:
        id:
          type: integer
          format: int64
          example: 100
        fileName:
          type: string
          example: "TP1_Rendus.zip"
        pathStorage:
          type: string
          description: Chemin de stockage du ZIP original.
          example: "/data/courses/61-13/TP1/TP1_Rendus.zip"
        pathFileStructured:
          type: string
          description: Chemin du ZIP restructuré.
          example: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
      required: ["fileName", "pathStorage"]
      example:
        id: 100
        fileName: "TP1_Rendus.zip"
        pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
        pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
    TPDto:
      type: object
      description: Représentation d’un TP pour un cours.
      properties:
        id:
          type: integer
          format: int64
          example: 10
        "no":
          type: integer
          example: 1
        courseId:
          type: integer
          format: int64
          example: 1
        submission:
          $ref: '#/components/schemas/SubmissionDto'
        statusStudents:
          type: array
          items:
            $ref: '#/components/schemas/TPStatusDto'
      required: ["no", "courseId"]
      example:
        id: 10
        no: 1
        courseId: 1
        submission:
          id: 100
          fileName: "TP1_Rendus.zip"
          pathStorage: "/data/courses/61-13/TP1/TP1_Rendus.zip"
          pathFileStructured: "/data/courses/61-13/TP1/TP1_RenduRestructuration.zip"
        statusStudents:
          - id: 1000
            studentId: 300
            tpId: 10
            studentSubmission: true
    TPStatusDto:
      type: object
      description: Statut d’un étudiant pour un TP.
      properties:
        id:
          type: integer
          format: int64
          example: 1000
        studentId:
          type: integer
          format: int64
          example: 300
        tpId:
          type: integer
          format: int64
          example: 10
        studentSubmission:
          type: boolean
          example: true
      required: ["studentId", "tpId", "studentSubmission"]
      example:
        id: 1000
        studentId: 300
        tpId: 10
        studentSubmission: true